refactor: move `settings_manager` to core
