+++ gui/update_dialog.py
@@ -15,7 +15,7 @@ from PyQt6.QtWidgets import (
 )
 
 from core.auto_updater import perform_updates
-from gui.settings_manager import SettingsManager
+from core.settings import SettingsManager
 
 log = logging.getLogger()
